chore: Cleanup qis-compiler file org
